streamline CLI serve command
